feat: commit functioning
