+++ src/app/screens/Calculator/styles.module.scss
