+++ src/app/screens/Record/styles.module.scss
@@ -89,8 +89,8 @@
 
 .clearButton {
   @extend%recordButtons;
-    text-align: left;
-    padding-left: 15px;
+  text-align: left;
+  padding-left: 15px;
 }
 
 .recToCalcButton {
