+++ src/app/screens/Survey/form/validate.js
@@ -1,15 +1,10 @@
-const validate = values => {
+export const validate = values => {
   const errors = {};
   const requiredFields = ['name', 'input', 'phone'];
   requiredFields.forEach(field => {
     if (!values[field]) {
-      errors[field] = 'Required';
+      errors[field] = 'Campo obligatorio';
     }
   });
-  if (values.email && !/^[0-9]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.email)) {
-    errors.email = 'Invalid email address';
-  }
   return errors;
 };
-
-export default validate;
