+++ src/app/screens/Survey/index.js
@@ -1,18 +1,13 @@
-import { arrayOf, elementType } from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 
-import OperationActions from 'redux/operations/actions';
-
-import RecordLayout from './layout';
+import SurveyLayout from './layout';
 import styles from './styles.module.scss';
 
-const Survey = () => {
-  return (
-    <div className={styles.container}>
-      <RecordLayout />
-    </div>
-  );
-};
+const Survey = () => (
+  <div className={styles.container}>
+    <SurveyLayout />
+  </div>
+);
 
 export default connect()(Survey);
