+++ src/app/screens/Survey/styles.module.scss
@@ -14,66 +14,66 @@
 }
 
 .surveyTitle {
-    align-content: center;
-    align-items: flex-start;
-    background-color: $gray;
-    color: $disabled-gray;
-    display: flex;
-    flex-direction: column;
-    font-size: 25px;
-    font-weight: 300;
-    height: 70px;
-    justify-content: center;
-    padding-left: 12px;
-  }
+  align-content: center;
+  align-items: flex-start;
+  background-color: $gray;
+  color: $disabled-gray;
+  display: flex;
+  flex-direction: column;
+  font-size: 25px;
+  font-weight: 300;
+  height: 70px;
+  justify-content: center;
+  padding-left: 12px;
+}
 
 .formContainer {
-    background-color: $plain-white;
+  background-color: $plain-white;
 }
 
 .formField {
-    width: 80%; 
-    margin: 15px;     
+  width: 80%;
+  margin: 15px;
 }
 
-  .surveyButtons {
-    align-items: stretch;
-    background-color: $gray;
-    border: 0;
-    color: $disabled-gray;
-    display: flex;
-    font-size: 16px;
-    font-weight: 300;
-    justify-content: space-between;
-    width: 100%;
-  }
-  
-  %surveyButtons {
-    appearance: none;
-    background-color: $gray;
-    border: 0;
-    border-radius: 0;
-    color: $disabled-gray;
-    font-size: 16px;
-    font-weight: 300;
-    padding: 10px;
-    text-align: left;
-    width: 50%;
-  
-    &:hover {
-      background-color: $black;
-    }
-  }
-  
-  .submitButton {
-    @extend%surveyButtons;
-    text-align: left;
-    padding-left: 15px;
-  }
+.surveyButtons {
+  align-items: stretch;
+  background-color: $gray;
+  border: 0;
+  color: $disabled-gray;
+  display: flex;
+  font-size: 16px;
+  font-weight: 300;
+  justify-content: space-between;
+  width: 100%;
+}
 
-  .survToCalcButton {
-    @extend%surveyButtons;
-      text-align: right;
-      padding-right: 15px;
+%surveyButton {
+  appearance: none;
+  background-color: $gray;
+  border: 0;
+  border-radius: 0;
+  color: $disabled-gray;
+  font-size: 16px;
+  font-weight: 300;
+  padding: 10px;
+  text-align: left;
+  width: 50%;
+
+  &:hover {
+    background-color: $black;
   }
-  
+}
+
+.submitButton {
+  @extend%surveybutton !optional;
+  text-align: left;
+  padding-left: 15px;
+}
+
+.survToCalcButton {
+  @extend%surveybutton !optional;
+  text-align: right;
+  padding-right: 15px;
+}
+
